pyproject.toml
[project]
name = "llm-security-benchmarking"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "dvc>=3.0",
    "pyyaml>=6.0",
]
